remove Concurrent something class
